Move draft components to app module
