+++ console/frontend/src/app/app/draft.ts
