+++ console/frontend/src/app/new-draft-editor/new-draft-editor.component.ts
@@ -9,7 +9,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 
-import { NewDraftForm } from '../new-draft-form';
+import { NewDraftForm } from './new-draft-form';
 
 @Component({
     selector: 'app-new-draft-editor',
