+++ console/frontend/src/app/new-draft-editor/new-draft-form.ts
